Added network test and a base test class
