+++ systest/lib/__init__.py
@@ -4,7 +4,7 @@ systest.lib package
 common libraries for system tests
 '''
 
-__all__ = ['data']
+__all__ = ['data', 'netem', 'mp_helper']
 
 import importlib
 for m in __all__:
